Reduce duplicate code between appearancesListDialogSlices/Sagas
