+++ premiser-ui/src/components/mediaExcerptApparitionsDialog/mediaExcerptApparitionsDialogSlice.ts
@@ -1,112 +1,13 @@
-import { ActionCreator, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { call, delay, put, race, take, takeEvery } from "typed-redux-saga";
-import { normalize } from "normalizr";
-import { LOCATION_CHANGE } from "connected-react-router";
+import { api } from "@/apiActions";
+import {
+  createAppearancesDialogListSaga,
+  createAppearancesDialogListSlice,
+} from "../appearanceListDialog/appearancesListDialogSlices";
 
-import { EntityId } from "howdju-common";
-
-import config from "@/config";
-import { RootState } from "@/setupStore";
-import { api, ApiResponseActionMeta } from "@/apiActions";
-
-const initialState = {
-  isDialogVisible: false,
-  appearanceIds: [] as EntityId[],
-  fetchError: undefined as string | undefined,
-};
-
-const slice = createSlice({
-  name: "MediaExcerptApparitionsDialog",
-  initialState,
-  reducers: {
-    showDialog: (state, _action: PayloadAction<EntityId>) => {
-      state.isDialogVisible = true;
-      state.fetchError = undefined;
-    },
-    hideDialog: (state) => {
-      state.isDialogVisible = false;
-    },
-    fetchAppearancesSuccess: (
-      state,
-      { payload: appearanceIds }: PayloadAction<EntityId[]>
-    ) => {
-      state.appearanceIds = appearanceIds;
-      state.fetchError = undefined;
-    },
-    fetchAppearancesFailure: (state) => {
-      state.fetchError = "Error fetching appearances";
-    },
-  },
-  extraReducers(builder) {
-    builder.addCase(LOCATION_CHANGE, (state) => {
-      // If the user navigates, hide the dialog.
-      // I.e. if the user clicks on an appearance link in the dialog, hide the dialog.
-      state.isDialogVisible = false;
-    });
-  },
-});
+const slice = createAppearancesDialogListSlice("PropositionAppearancesDialog");
 
 export default slice.actions;
 export const mediaExcerptApparitionsDialog = slice.reducer;
 
-export const selectIsDialogVisible = (state: RootState) =>
-  state.mediaExcerptApparitionsDialog.isDialogVisible;
-
-export function* mediaExcerptApparitionsDialogSaga() {
-  yield takeEvery(
-    slice.actions.showDialog,
-    // TODO(525) factor out the duplicate logic in these sagas. In particular, this slice is nearly
-    // identical to propositionAppearancesDialogSlice.
-    function* ({ payload: mediaExcerptId }) {
-      const {
-        payload: {
-          fetchInit: { requestId },
-        },
-      } = yield put(api.fetchMediaExcerptApparitions(mediaExcerptId));
-      const {
-        responseAction,
-        timeout,
-      }: {
-        responseAction: ReturnType<
-          typeof api.fetchMediaExcerptApparitions.response
-        >;
-        timeout: true;
-      } = yield race({
-        responseAction: call(
-          getResponseAction,
-          requestId,
-          api.fetchMediaExcerptApparitions.response
-        ),
-        timeout: delay(config.apiTimeoutMs),
-      });
-      if (responseAction) {
-        if (!responseAction.error) {
-          const { result } = normalize(
-            responseAction.payload,
-            responseAction.meta.normalizationSchema
-          );
-          const appearanceIds = result.appearances;
-
-          yield put(slice.actions.fetchAppearancesSuccess(appearanceIds));
-        } else {
-          yield put(slice.actions.fetchAppearancesFailure());
-        }
-      } else if (timeout) {
-        yield put(slice.actions.fetchAppearancesFailure());
-      } else {
-        yield put(slice.actions.fetchAppearancesFailure());
-      }
-    }
-  );
-}
-
-function* getResponseAction<
-  T extends PayloadAction<any, string, ApiResponseActionMeta<any, any>>
->(requestId: string, responseActionCreator: ActionCreator<any>) {
-  while (true) {
-    const resultAction: T = yield take(responseActionCreator);
-    if (resultAction.meta.requestId === requestId) {
-      return resultAction;
-    }
-  }
-}
+export const mediaExcerptApparitionsDialogSaga =
+  createAppearancesDialogListSaga(slice, api.fetchPropositionAppearances);
